Make the ticks slanted and display all ticks
