+++ app/pathogen/arbovirus/analyze/recharts.tsx
@@ -325,7 +325,7 @@ type AgeGroup =
   | "Seniors (65+ years)"
   | "Multiple groups";
 
-function CustomizedWHORegionTick(props: any) {
+export function SlantedTick(props: any) {
   const { x, y, payload } = props;
 
   return (
@@ -475,7 +475,7 @@ export function MedianSeroPrevByWHOregion() {
                 <XAxis
                   dataKey="region"
                   interval={0}
-                  tick={<CustomizedWHORegionTick />}
+                  tick={<SlantedTick />}
                 />
                 <YAxis
                   domain={[0, 100]}
@@ -660,7 +660,7 @@ export function MedianSeroPrevByWHOregionAndAgeGroup() {
                   <XAxis
                     dataKey="region"
                     interval={0}
-                    tick={<CustomizedWHORegionTick />}
+                    tick={<SlantedTick />}
                   />
                   <YAxis
                     domain={[0, 100]}
